feat(landing-page): completed multilanguage support
