--- src/utils/util-constants.ts
@@ -1,183 +1,0 @@
-export const selectLanguageSettings = [
-  {
-    text: "Node JS",
-    key: "nodejs",
-    value: "nodejs",
-  },
-  {
-    text: "Java",
-    key: "java",
-    value: "java",
-  },
-  {
-    text: "Python 3",
-    key: "python3",
-    value: "python3",
-  },
-  {
-    text: "Python 2",
-    key: "python2",
-    value: "python2",
-  },
-  {
-    text: "C",
-    key: "c",
-    value: "c",
-  },
-  {
-    text: "C++",
-    key: "cpp",
-    value: "cpp",
-  },
-  {
-    text: "PHP",
-    key: "php",
-    value: "php",
-  },
-  {
-    text: "Ruby",
-    key: "ruby",
-    value: "ruby",
-  },
-  {
-    text: "Go Lang",
-    key: "go",
-    value: "go",
-  },
-  {
-    text: "Bash",
-    key: "bash",
-    value: "bash",
-  },
-  {
-    text: "SQL",
-    key: "sql",
-    value: "sql",
-  },
-  {
-    text: "C#",
-    key: "csharp",
-    value: "csharp",
-  },
-  {
-    text: "Dart",
-    key: "dart",
-    value: "dart",
-  },
-];
-
-export const coreLanguageSettings = [
-  {
-    key: "nodejs",
-    versionIndex: "4",
-    mode: "javascript",
-  },
-  {
-    key: "java",
-    versionIndex: "4",
-    mode: "java",
-  },
-  {
-    key: "python3",
-    versionIndex: "4",
-    mode: "python",
-  },
-  {
-    key: "python2",
-    versionIndex: "3",
-    mode: "python",
-  },
-  {
-    key: "c",
-    versionIndex: "5",
-    mode: "c_cpp",
-  },
-  {
-    key: "cpp",
-    versionIndex: "5",
-    mode: "c_cpp",
-  },
-  {
-    key: "php",
-    versionIndex: "4",
-    mode: "php",
-  },
-  {
-    key: "ruby",
-    versionIndex: "4",
-    mode: "ruby",
-  },
-  {
-    key: "go",
-    versionIndex: "4",
-    mode: "golang",
-  },
-  {
-    key: "bash",
-    versionIndex: "4",
-    mode: "textile",
-  },
-  {
-    key: "sql",
-    versionIndex: "4",
-    mode: "sql",
-  },
-  {
-    key: "csharp",
-    versionIndex: "4",
-    mode: "csharp",
-  },
-  {
-    key: "dart",
-    versionIndex: "4",
-    mode: "dart",
-  },
-];
-
-export const themeSettings = [
-  {
-    text: "Tomorrow",
-    key: "tomorrow",
-    value: "tomorrow",
-  },
-  {
-    text: "Monokai",
-    key: "monokai",
-    value: "monokai",
-  },
-  {
-    text: "GitHub",
-    key: "github",
-    value: "github",
-  },
-  {
-    text: "Kuroir",
-    key: "kuroir",
-    value: "kuroir",
-  },
-  {
-    text: "Twilight",
-    key: "twilight",
-    value: "twilight",
-  },
-  {
-    text: "XCode",
-    key: "xcode",
-    value: "xcode",
-  },
-  {
-    text: "Text Mate",
-    key: "textmate",
-    value: "textmate",
-  },
-  {
-    text: "Solarized Dark",
-    key: "solarized_dark",
-    value: "solarized_dark",
-  },
-  {
-    text: "Solarized Light",
-    key: "solarized_light",
-    value: "solarized_light",
-  },
-];
